examples/multimodal_crawler/requirements.txt
firecrawl==0.0.1  # all libraries are most up to date as of 2025-01-01
pydantic==2.5.2
python-dotenv==1.0.0